[wip] Change mqtt_subscription to disc_only_copies
Fixes quadratic startup time in number of subscriptions for a single user, due to
ets bag tables not coping well with keys having lots of values.

This is a quick hack to demonstrate the fix. A proper fix should change
ets/mnesia calls to go through kvs, which would require extending the kvs API.